packages/range-safety/pyproject.toml
[project]
name = "hangar-range-safety"
version = "0.1.0"
description = "Plan validation and post-run assertions for MDAO workflows"
requires-python = ">=3.11"
dependencies = [
    "hangar-sdk[all]",
    "hangar-omd",
    "jsonschema>=4.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
range-safety = "hangar.range_safety.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hangar"]

[tool.uv.sources]
hangar-sdk = { workspace = true }
hangar-omd = { workspace = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
